add layer application and backendApi

+++ eProjectNetCore.Data/Entites/Product.cs
@@ -8,7 +8,7 @@ namespace eProjectNetCore.Data.Entites
     {
         public int Id { get; set; }
         public decimal Price { get; set; }
-        public int OriginalPrice { get; set; }
+        public decimal OriginalPrice { get; set; }
         public int Stock { get; set; }
         public int ViewCount { get; set; }
         public DateTime DateCreated { get; set; }
